Refactor session and shell directories
