+++ internal/session/getSessionHistoryLines.go
