+++ internal/session/getSessionLogs.go
